Extract dialog code to separate module
