+++ ckChangeLog-dialog/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest package="de.cketti.library.changelog.dialog">
+
+    <application/>
+
+</manifest>
