+++ ckChangeLog-dialog/src/main/res/values-cs/strings.xml
