+++ ckChangeLog-dialog/src/main/res/values-de/strings.xml
