+++ ckChangeLog-dialog/src/main/res/values-el/strings.xml
