+++ ckChangeLog-dialog/src/main/res/values-es/strings.xml
