+++ ckChangeLog-dialog/src/main/res/values-pl/strings.xml
