+++ ckChangeLog-dialog/src/main/res/values-ru/strings.xml
