+++ ckChangeLog-dialog/src/main/res/values-sk/strings.xml
