+++ ckChangeLog-dialog/src/main/res/values-uk/strings.xml
