+++ ckChangeLog-dialog/src/main/res/values/strings.xml
