+++ sample/src/main/java/de/cketti/sample/changelog/MainActivity.java
@@ -1,6 +1,7 @@
 package de.cketti.sample.changelog;
 
-import de.cketti.library.changelog.ChangeLog;
+import de.cketti.library.changelog.dialog.DialogChangeLog;
+
 import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
@@ -10,13 +11,16 @@ import android.view.MenuItem;
 
 
 public class MainActivity extends FragmentActivity {
+    private static final String DARK_THEME_CSS = "body { color: #ffffff; background-color: #282828; }" + "\n" +
+            DialogChangeLog.DEFAULT_CSS;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        ChangeLog cl = new ChangeLog(this);
+        DialogChangeLog cl = DialogChangeLog.newInstance(this);
         if (cl.isFirstRun()) {
             cl.getLogDialog().show();
         }
@@ -32,11 +36,11 @@ public class MainActivity extends FragmentActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case R.id.menu_whats_new: {
-                new DarkThemeChangeLog(this).getLogDialog().show();
+                createDarkThemeChangeLog(this).getLogDialog().show();
                 break;
             }
             case R.id.menu_full_changelog: {
-                new ChangeLog(this).getFullLogDialog().show();
+                DialogChangeLog.newInstance(this).getFullLogDialog().show();
                 break;
             }
         }
@@ -44,15 +48,8 @@ public class MainActivity extends FragmentActivity {
         return true;
     }
 
-    /**
-     * Example that shows how to create a themed dialog.
-     */
-    public static class DarkThemeChangeLog extends ChangeLog {
-        public static final String DARK_THEME_CSS =
-                "body { color: #ffffff; background-color: #282828; }" + "\n" + DEFAULT_CSS;
-
-        public DarkThemeChangeLog(Context context) {
-            super(new ContextThemeWrapper(context, R.style.DarkTheme), DARK_THEME_CSS);
-        }
+    private static DialogChangeLog createDarkThemeChangeLog(Context context) {
+        ContextThemeWrapper themedContext = new ContextThemeWrapper(context, R.style.DarkTheme);
+        return DialogChangeLog.newInstance(themedContext, DARK_THEME_CSS);
     }
 }
